Change WGM and WGMXCC to int, matching sizemapping

+++ shared/origami/include/origami/origami.hpp
@@ -35,7 +35,7 @@ prediction_result_t select_config(const problem_t& problem,
  * @param wgms List of possible workgroup-mappings.
  * @return std::tuple<size_t, size_t>
  */
-std::tuple<size_t, size_t> select_workgroup_mapping(const problem_t& problem,
+std::tuple<int, int> select_workgroup_mapping(const problem_t& problem,
                                                     const hardware_t& hardware,
                                                     const config_t& config,
                                                     size_t skGrid);
